Move TodoApp previews to compose-ui module
