+++ examples/todoapp/common/compose-ui/src/desktopMain/kotlin/example/todo/common/ui/TodoEditPreview.kt
@@ -1,8 +1,4 @@
-/*
- * Should be in the compose-ui module, see https://github.com/JetBrains/compose-jb/issues/908
- */
-
-package example.todo.desktop
+package example.todo.common.ui
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
@@ -10,7 +6,6 @@ import com.arkivanov.decompose.value.MutableValue
 import com.arkivanov.decompose.value.Value
 import example.todo.common.edit.TodoEdit
 import example.todo.common.edit.TodoEdit.Model
-import example.todo.common.ui.TodoEditContent
 
 @Composable
 @Preview
