+++ examples/todoapp/desktop/build.gradle.kts
@@ -25,11 +25,6 @@ kotlin {
                 implementation(Deps.ArkIvanov.MVIKotlin.mvikotlinMain)
                 implementation(Deps.Badoo.Reaktive.reaktive)
                 implementation(Deps.Badoo.Reaktive.coroutinesInterop)
-
-                // The dependencies below are required only for previews, see https://github.com/JetBrains/compose-jb/issues/908
-                implementation(compose.uiTooling)
-                implementation(project(":common:edit"))
-                implementation(project(":common:main"))
             }
         }
     }
